исправенна ошибка с отображением большой картинки
